Make content streams work with streams from `readable-stream`
`readable-stream` goes to some effort to make sure that all of its streams
inherit from `stream.Stream` ultimately.  What it doesn't do is ensure that
any of the various kinds of streams are `instanceof` the system streams.
So an `instanceof` check on a `stream.Readable` will fail for those
constructed with `readable-stream`.